User model: add FlowField and CredentialsField
We originally needed FlowModel and CredentialsModel because we were using
the built-in Django User model.

We wanted to store Google OAuth2 credentials in our db, and we needed a
way to relate those credentials to Users, which we accomplished via the
foreign keyed models above.

Now that we have our own custom Raven model, we can simply create two
new fields directly, and just store the credentials and flow objects
in there.

This should also make our future South migrations quite a bit simpler.
(note, they are currently turned off, and will need to be turned back on
after this commit)