requirements.txt: deploy libgreader from a tarball
We do this instead of grabbing directly from github. This is required
because we have a private copy of libgreader in our github, which is only
accessible on heroku either via our private SSH key or by encoding a
personal github password in requirements.txt, both of which are sucky
ideas.

We can grab this file from my web host for now. Once we get S3 setup,
we should grab it from there instead.