static: add SpecialElite font
This will be required because Google's web fonts are served via http,
and not https. Most modern browsers will refuse to load mixed content.

Therefore, we add the font to our own repo (it is Apache 2.0 licensed)
and adjust our style sheets accordingly.

Additionally, this font will now be copied to Amazon S3 when we do an
s3cmd sync, and served along with the rest of our static assets.