raven/models: add Feed as ForeignKey in UserFeedItem
A small tweak for future performance boost. We will want the db to
index the tuple of (user, feed, read, items) to quickly pull out all
the unread items in a feed.

So add the relationship, and create an index.

(There is a stray fix to an unrelated resource test in this commit, to
test on a blog that actually has updated feed items.)