usher/models/Feed: require explicit call to update() on new feeds
Previously, calling save() on a new Feed would schedule an UpdateFeedTask()
which would eventually call update(). This was a handy side-effect, but
it actually makes importing feeds from libgreader trickier, because in
the import case, we don't want to use feedparser for the import; we want
to use libgreader directly.