raven/models/FeedItem: add the timestamp to our GUID calculation
We need this because we've discovered some malformed feeds with duplicate
entries in the feed that hash to the same value. This happens typically
when a blogging platform deletes old entries and then re-publishes them for
some unknown reason.

That scenario is a valid thing to happen, but the problem is all the fields
feeding into our hash remain the same, except for the publishing time. So
we need to additionally hash on that field to prevent a database uniqueness
constraint validation error.

Note: in addition to the new test, I also ran a manual one-time test by
hand both before this commit and then after this commit. The manual test
patch follows below. The reason it is manual is because hard coding the
number of expected FeedItem objects will change over time. I checked that
both before and after the commit, total_items was 1596, which proves that
hashing on published time does not spuriously allow duplicate items into
the database.

--- a/raven/tests/test_tasks.py
+++ b/raven/tests/test_tasks.py
@@ -7,7 +7,7 @@ from django.test import TestCase
 from django.test.utils import override_settings

 from raven import tasks
-from raven.models import Feed
+from raven.models import Feed, FeedItem
 from raven.test_utils import network_available

 THIS_DIR = os.path.dirname(__file__)
@@ -111,3 +111,7 @@ class SyncFromReaderAPITaskTest(TestCase):
         total_feeds = Feed.objects.all().count()
         owner = User.objects.get(pk=owner.pk)
         self.assertEqual(owner.feeds.count(), total_feeds-1)
+
+        total_items = FeedItem.objects.all().count()
+        print total_items