usher/models/UserFeedItem: hook into FeedItem post_save signal
We should never have to manually create an entry in the UserFeedItem
join table because we will surely forget to do this at some point.

Instead, every time a new FeedItem is created, guarantee that all users
subscribed to the feed see the new item.

This is a noisy diff because I had to move the definitions of UserFeed
and UserFeedItem below the definitions of Feed and FeedItem classes,
because python doesn't seem to have forward declarations. Yes, C is still
sometimes superior.