raven/models/Feed: hash on reader_guid too
When we import from Google Reader, we want to ensure that we hash on
reader_guid, because a feed item may not be unique otherwise. This can
happen if the feed hiccups for whatever reason, and Reader sees two
copies of it. The good news is that Reader's guid will still be unique.