raven/models/Feed: differentiate between a link and a site
link is the RSS link we want to parse
site is the real web page that the RSS belongs to

Note that site is not required. Reader can return to us a valid link
without an associated site, and we really only care about the rss link,
so make site optional.