raven/models/Feed: add support of autodiscovering RSS feeds
In our UI, we will want to make it possible for the reader to just enter
a web site, say 'http://boingboing.net' which will subscribe to BB's RSS
feed.

We do not want the user to have to manually enter in
'http://boingboing.net/atom.xml' because that is a discovery nightmare.

So add an autodiscover interface, implemented via feedfinder.

Also, it feels very poignant that we've incorporated Aaron Swartz's code
into our machinery. We should seriously consider tithing some of our profit
to the EFF or similar...