usher/views: provide UI to import Google Takeout data
This is a good first cut and does work, but should be beefed up a bit.
In particular, it would be better to have a model handle the upload so
we can have an audit trail for users, as well as provide a history of
successful imports.

We also need to make the dashboard more ajaxy somehow.