usher/views: if user tries to go to /usher directly, redirect to dashboard
We don't really support an /usher endpoint, but if an annoying user tries
to go there, let's send them to their dashboard.

If they try and go to /usher before even logging on, then we simply ask
for their google auth credentials, and send them to the default home view.