Get pagination working.
The current idea is that the feeds will continue to fetch until there are no
more. items, on the other hand, will need an explicit call to getNext(), which
we'll hook up as an infinite scrolling thing.

This has been tested on 245 feeds and 18,497 items (seriously, Reader isn't
actually pruning stuff, I guess).