usher: fix local dev new user creation
As we no longer have a debug stripe instance, we can't use a test credit
card anymore. Unfortunately, this also means when you try and create a new
user, it bombs out because you won't be in the production stripe db.

Therefore, we:

	- always return true in is_customer()
	- but the sign_up() view only redirects to / if not DEBUG

When creating a new user in local dev mode, when prompted for the credit
card, just cancel out and navigate to / manually.