raven/models: do not use publish date in GUID calculation
There are too many broken feeds out there that do not publish dates.

Since we stick utcnow() into the 'published' field on those crap feeds,
we don't want to use that time to calculate the GUID, as utcnow() will
be different next time we pull the feed and we'll end up with a duplicate
FeedItem in our database.

Ugh.

Fixes #78