usher/views: only SyncFromReader a single time
Previously we only sync'ed from Reader a single time because we did that
step in the 'ask for credit card' page.

When we removed the credit card page, we moved the sync task into the
normal sign-in process.... except this will cause us to sync *every* time
the user logs out and logs in.

Oops.

Instead, check for the sync_task_id and if we don't find one, then
kick off the task.