raven/models: use get_or_create() everywhere
Hopefully this kills our db integrity errors once and for all.

The key learning out of all of this was that validate_unique() and
get_or_create() work when you only pass in the fields that are constrained
as unique.

For example:

	unique_together = (feed, guid)

Now imagine you do something like the following:

	item, new = FeedItem.objects.get_or_create(feed=feed, guid=guid,
						   extra=extra)

This will actually fail, because the presence of 'extra' will screw up
the uniqueness test.

The first thing to do would be leave off the 'extra' field. But what if
'extra' is constrained to be !null? You'll fail for a different reason.

The final answer is to use the 'default' dictionary to satisfy all
non-null constraints, but still allow the uniqueness checks to work properly.