db: data migration, update values of FETCH_DEFAULT
We made a commit where FETCH_DEFAULT was 60 minutes.

We made a later commith where we changed FETCH_DEFAULT to 30.

Our query looks like this:

  for freq in [Feed.FETCH_FAST, Feed.FETCH_DEFAULT, Feed.FETCH_SLOW]:
      age = datetime.utcnow() - timedelta(minutes=freq)
      feeds = Feed.objects.filter(last_fetched__lt=age,
			          fetch_frequency=freq)[:self.SLICE_SIZE]

Notice how we will search for fetch_frequency=freq. Well, if freq == 30
but all the items in the db have the value of 60, then our queryset will
always be empty.

Thus, migrate all the 60 => 30.