raven/models/feed: fix one last splat
We are still seeing a traceback:

  [2013-06-09 18:45:25,101: ERROR/MainProcess] Task
raven.tasks.SyncFromReaderAPITask[dd390cca-1eae-474b-bc77-736ea4873b63] raised
exception: UnicodeEncodeError('ascii', u'Manorexia \u2013 Eating Disorders in
Men', 10, 11, 'ordinal not in range(128)') 

  Traceback (most recent call last): 
    File "/app/.heroku/python/lib/python2.7/site-packages/celery/task/trace.py", line 228, in trace_task 
      R = retval = fun(*args, **kwargs) 
    File "/app/.heroku/python/lib/python2.7/site-packages/newrelic-1.11.0.55/newrelic/api/object_wrapper.py", line 216, in __call__ 
      self._nr_instance, args, kwargs) 
    File "/app/.heroku/python/lib/python2.7/site-packages/newrelic-1.11.0.55/newrelic/hooks/application_celery.py", line 62, in wrapper  
      return wrapped(*args, **kwargs) 
    File "/app/.heroku/python/lib/python2.7/site-packages/celery/task/trace.py", line 415, in __protected_call__ 
      return self.run(*args, **kwargs) 
    File "/app/raven/tasks.py", line 314, in run 
      _new_user_item(user, feed, e) 
    File "/app/raven/tasks.py", line 44, in _new_user_item 
      item.guid = item.calculate_guid() 
    File "/app/raven/models.py", line 380, in calculate_guid 
      guid.update(self.title) 

  UnicodeEncodeError: 'ascii' codec can't encode character u'\u2013' in
position 10: ordinal not in range(128) 

This is extremely strange, because we *should* be encoding the title
as per:
	tmp.title = HTMLParser().unescape(entry.title).encode('utf-8')

But clearly something went awry.

Let's just go back to the old way of doing things, where we assume
all the data going into calculate_guid() must be explicitly encoded in
utf-8.

But for fun, store the title in the db as utf-8 too.