raven/models/feeditem: add some new database indexes
New Relic is telling us the following 2 queries are slow:

	SELECT "raven_feeditem"."id", "raven_feeditem"."feed_id",
	"raven_feeditem"."description", "raven_feeditem"."link",
	"raven_feeditem"."title", "raven_feeditem"."guid",
	"raven_feeditem"."atom_id", "raven_feeditem"."reader_guid",
	"raven_feeditem"."published" FROM "raven_feeditem" WHERE
	("raven_feeditem"."guid" = %s AND "raven_feeditem"."feed_id" = %s )

and:

	SELECT COUNT(*) FROM "raven_feeditem" WHERE ("raven_feeditem"."feed_id"
	= %s AND "raven_feeditem"."published" > %s )

So let's add an index for those fields, together.

This requires a new version of south, 0.8.1 that supports the index_together
meta attribute. It also turns out that we never actually indexed the fields
we thought we were indexing before.

Ouch.