raven/models: optimize FeedItem._get_or_create()
This function is showing up as an extremely slow SQL trace because one
of the things it does is search the db for the 'link' column, and we don't
have an index for it.

Add an index, but instead of indexing on 'link', add a new column named
'link_hash' and index that instead, as 99% of all links will start with
'http://' which will index poorly.