raven: proper calculation of last_fetched
We previously had a hack of touching every feed when it got added to
the fetch queue, to control how often feeds got entered into the queue.
This scheme didn't work, and we often ended up fetching the same feed
multiple times within a few seconds when attempting to clear out the
queue.

Instead:
	- update Feed.last_fetched only after a successful fetch.
	- have Feed.update() compare last_fetched vs fetch_frequency and
	  return if it's too soon